chore(dal): namespace renaming
