+++ OnlineGameStore.DAL.Tests/RepositoryCreators/GameRepositoryCreator.cs
@@ -1,5 +1,5 @@
 using OnlineGameStore.DAL.Repositories;
 
-namespace OnlineGameStore.DAL.Tests;
+namespace OnlineGameStore.DAL.Tests.RepositoryCreators;
 
 public class GameRepositoryCreator : RepositoryCreator<GameRepository> { }
